update state when log in and go back to home page
