+++ src/components/App.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { Route, Switch, BrowserRouter } from 'react-router-dom';
 import Navbar from './Navbar';
 import Home from './Home';
@@ -8,8 +8,17 @@ import UserLogin from './UserLogin';
 import UserRegister from './UserRegister';
 import '../styles/App.css';
 
-function App() {
+const initialState = {
+  user: {
+    loggedIn: false,
+    userName: '',
+    translateFrom: '',
+    translateTo: '',
+  },
+};
 
+function App() {
+  const [user, setUser] = useState(initialState.user);
   return (
     <BrowserRouter>
       <div className="App">
@@ -17,10 +26,9 @@ function App() {
         <Switch>
           <Route exact path="/profile" component={Profile} />
           <Route path="/learn" component={Learn} />
-          <Route path="/login" component={UserLogin} />
-          <Route path="/register" component={UserRegister} />
-          <Route path="/" component={Home} />
-
+          <Route exact path="/login" render={() => <UserLogin userState={user} setUserState={setUser} />} />
+          <Route exact path="/register" render={() => <UserRegister userState={user} setUserState={setUser} />} />
+          <Route exact path="/" render={() => <Home userState={user} setUserState={setUser} />} />
         </Switch>
       </div>
     </BrowserRouter>
